CD-154: Changes according to review
