+++ source/com.unity.cluster-display/Runtime/ClusterSync.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Runtime.InteropServices;
 using UnityEngine;
 
 #if UNITY_EDITOR
@@ -150,10 +149,12 @@ namespace Unity.ClusterDisplay
                 return;
             }
 
+#if UNITY_STANDALONE_WIN
             if (NodeRole != NodeRole.Unassigned)
             {
-                CheckNodeSetup();
+                ClusterSyncWindowsHelpers.CheckNodeSetup();
             }
+#endif
 
             RegisterDelegates();
             ClusterSyncLooper.InjectSynchPointInPlayerLoop();
@@ -263,66 +264,6 @@ namespace Unity.ClusterDisplay
                 CleanUp();
                 return false;
             }
-        }
-
-#if UNITY_STANDALONE_WIN
-        [DllImport("advapi32.dll")]
-        static extern int RegOpenKeyEx(UIntPtr hKey, string subKey, int ulOptions, int samDesired, out UIntPtr hkResult);
-        [DllImport("advapi32.dll", SetLastError = true)]
-        static extern int RegCloseKey(UIntPtr hKey);
-        [DllImport("advapi32.dll", SetLastError = true)]
-        static extern uint RegQueryValueEx(UIntPtr hKey, string lpValueName, int lpReserved, out int lpType,
-            IntPtr lpData, ref int lpcbData);
-        static UIntPtr k_HKLM = new UIntPtr(0x80000002u);
-        const int k_ErrorSuccess = 0;
-        const int k_ValueKindDword = 4;
-        const int k_KeyReadAccess = 0x20019; // KEY_READ
-        const int k_KeyWOW64 = 0x0100; // KEY_WOW64_64KEY
-#endif
-
-        void CheckNodeSetup()
-        {
-#if UNITY_STANDALONE_WIN
-            UIntPtr keyHandle = UIntPtr.Zero;
-            IntPtr valueIntPtr = IntPtr.Zero;
-            try
-            {
-                if (RegOpenKeyEx(k_HKLM, @"SOFTWARE\Microsoft\Windows NT\CurrentVersion\Multimedia\SystemProfile",
-                        0, k_KeyReadAccess | k_KeyWOW64, out keyHandle) == k_ErrorSuccess)
-                {
-                    int valueKind;
-                    int valueSize = (int)Marshal.SizeOf<uint>();
-                    valueIntPtr = Marshal.AllocHGlobal(valueSize);
-                    uint ret = RegQueryValueEx(keyHandle, "NetworkThrottlingIndex", 0, out valueKind,
-                        valueIntPtr, ref valueSize);
-                    if (ret == k_ErrorSuccess && valueKind == k_ValueKindDword && valueSize == Marshal.SizeOf<uint>())
-                    {
-                        var networkThrottlingIndex = (uint)Marshal.ReadInt32(valueIntPtr);
-                        if (networkThrottlingIndex != uint.MaxValue)
-                        {
-                            ClusterDebug.LogWarning(
-                                "Multimedia Class Scheduler Service (MMCSS) NetworkThrottlingIndex not set to 0xFFFFFFFF which can increase network latency and so reduce framerate.");
-                        }
-                    }
-                    else
-                    {
-                        ClusterDebug.LogWarning(
-                            "Failed validation of Multimedia Class Scheduler Service (MMCSS) NetworkThrottlingIndex.  Most likely the key is missing which is the equivalent of the default behavior, consider fixing the registry key to improve network latency and at the same time framerate.");
-                    }
-                }
-            }
-            finally
-            {
-                if (keyHandle != UIntPtr.Zero)
-                {
-                    RegCloseKey(keyHandle);
-                }
-                if (valueIntPtr != IntPtr.Zero)
-                {
-                    Marshal.FreeHGlobal(valueIntPtr);
-                }
-            }
-#endif
         }
 
         public void CleanUp()
